% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adventofcode2021-package.R
\docType{package}
\name{adventofcode2021-package}
\alias{adventofcode2021}
\alias{adventofcode2021-package}
\title{adventofcode2021: Advent of Code 2021 Solutions in R}
\description{
Solutions to Advent of Code 2021 programming puzzles, written as an R package.
}
\references{
\url{https://adventofcode.com/2021}
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Selbosh/adventofcode2021/issues}
}

}
\author{
\strong{Maintainer}: David Antony Selby \email{david.selby@manchester.ac.uk} (\href{https://orcid.org/0000-0001-8026-5663}{ORCID})

Other contributors:
\itemize{
  \item Eric Wastl (Created puzzles and descriptions in the documentation.) [contributor]
}

}
\keyword{internal}
