% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{binary_to_int}
\alias{binary_to_int}
\title{Convert binary digits to integer}
\usage{
binary_to_int(x)
}
\arguments{
\item{x}{vector of binary digits}
}
\value{
integer
}
\description{
Convert binary digits to integer
}
