% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day01.R
\name{day01}
\alias{day01}
\alias{f01a}
\alias{f01b}
\alias{example_data_01}
\title{Day 01: Sonar Sweep}
\usage{
f01a(x)

f01b(x)

example_data_01(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f01a(x)} returns .... For Part Two,
\code{f01b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/1}{Sonar Sweep}
}
\examples{
f01a(example_data_01())
f01b(example_data_01())
}
