% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{day03}
\alias{day03}
\alias{f03a}
\alias{f03b}
\alias{example_data_03}
\title{Day 03: Binary Diagnostic}
\usage{
f03a(x)

f03b(x)

example_data_03(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f03a(x)} returns .... For Part Two,
\code{f03b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/3}{Binary Diagnostic}
}
\examples{
f03a(example_data_03())
f03b(example_data_03())
}
