% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04}
\alias{day04}
\alias{f04a}
\alias{f04b}
\alias{get_numbers}
\alias{get_cards}
\title{Day 04: Giant Squid}
\usage{
f04a(numbers, cards)

f04b(numbers, cards)

get_numbers()

get_cards()
}
\arguments{
\item{numbers}{Numbers called so far}

\item{cards}{List of bingo cards}
}
\value{
For Part One, \code{f04a(x)} returns .... For Part Two,
\code{f04b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/4}{Giant Squid}
}
\details{
\strong{Part One}

You\'re already almost 1.5km (almost a mile) below the surface of the
ocean, already so deep that you can\'t see any sunlight. What you \emph{can}
see, however, is a giant squid that has attached itself to the outside
of your submarine.

Maybe it wants to play
\href{https://en.wikipedia.org/wiki/Bingo_(American_version)}{bingo}?

Bingo is played on a set of boards each consisting of a 5x5 grid of
numbers. Numbers are chosen at random, and the chosen number is \emph{marked}
on all boards on which it appears. (Numbers may not appear on all
boards.) If all numbers in any row or any column of a board are marked,
that board \emph{wins}. (Diagonals don\'t count.)

The submarine has a \emph{bingo subsystem} to help passengers (currently, you
and the giant squid) pass the time. It automatically generates a random
order in which to draw numbers and a random set of boards (your puzzle
input). For example:\preformatted{7,4,9,5,11,17,23,2,0,14,21,24,10,16,13,6,15,25,12,22,18,20,8,19,3,26,1

22 13 17 11  0
 8  2 23  4 24
21  9 14 16  7
 6 10  3 18  5
 1 12 20 15 19

 3 15  0  2 22
 9 18 13 17  5
19  8  7 25 23
20 11 10 24  4
14 21 16 12  6

14 21 17 24  4
10 16 15  9 19
18  8 23 26 20
22 11 13  6  5
 2  0 12  3  7
}

After the first five numbers are drawn (\code{7}, \code{4}, \code{9}, \code{5}, and \code{11}),
there are no winners, but the boards are marked as follows (shown here
adjacent to each other to save space):\preformatted{22 13 17 11  0         3 15  0  2 22        14 21 17 24  4
 8  2 23  4 24         9 18 13 17  5        10 16 15  9 19
21  9 14 16  7        19  8  7 25 23        18  8 23 26 20
 6 10  3 18  5        20 11 10 24  4        22 11 13  6  5
 1 12 20 15 19        14 21 16 12  6         2  0 12  3  7
}

After the next six numbers are drawn (\code{17}, \code{23}, \code{2}, \code{0}, \code{14}, and
\code{21}), there are still no winners:\preformatted{22 13 17 11  0         3 15  0  2 22        14 21 17 24  4
 8  2 23  4 24         9 18 13 17  5        10 16 15  9 19
21  9 14 16  7        19  8  7 25 23        18  8 23 26 20
 6 10  3 18  5        20 11 10 24  4        22 11 13  6  5
 1 12 20 15 19        14 21 16 12  6         2  0 12  3  7
}

Finally, \code{24} is drawn:\preformatted{22 13 17 11  0         3 15  0  2 22        14 21 17 24  4
 8  2 23  4 24         9 18 13 17  5        10 16 15  9 19
21  9 14 16  7        19  8  7 25 23        18  8 23 26 20
 6 10  3 18  5        20 11 10 24  4        22 11 13  6  5
 1 12 20 15 19        14 21 16 12  6         2  0 12  3  7
}

At this point, the third board \emph{wins} because it has at least one
complete row or column of marked numbers (in this case, the entire top
row is marked: \verb{14 21 17 24  4}).

The \emph{score} of the winning board can now be calculated. Start by finding
the \emph{sum of all unmarked numbers} on that board; in this case, the sum
is \code{188}. Then, multiply that sum by \emph{the number that was just called}
when the board won, \code{24}, to get the final score, \code{188 * 24 = 4512}.

To guarantee victory against the giant squid, figure out which board
will win first. \emph{What will your final score be if you choose that
board?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f04a(get_numbers(), get_cards())
f04a(get_numbers(), get_cards())
}
