% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day05.R
\name{day05}
\alias{day05}
\alias{f05}
\alias{example_data_05}
\title{Day 05: Hydrothermal Venture}
\usage{
f05(x, include_diag = FALSE)

example_data_05(example = 1)
}
\arguments{
\item{x}{some data}

\item{include_diag}{Include diagonal lines? Defaults to FALSE}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
Number of intersections
}
\description{
\href{https://adventofcode.com/2021/day/5}{Hydrothermal Venture}
}
\examples{
f05(example_data_05())
f05(example_data_05(), include_diag = TRUE)
}
