% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day06.R
\name{day06}
\alias{day06}
\alias{f06a}
\alias{f06b}
\title{Day 06: Lanternfish}
\usage{
f06a(x, days = 80)

f06b(x, days = 256)
}
\arguments{
\item{x}{some data}

\item{days}{number of days to simulate}
}
\description{
\href{https://adventofcode.com/2021/day/6}{Lanternfish}
}
\examples{
f06a(input_06())
f06b(input_06())
}
