% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{day07}
\alias{day07}
\alias{f07a}
\alias{f07b}
\alias{input_07}
\title{Day 07: The Treachery of Whales}
\usage{
f07a(x)

f07b(x)

input_07()
}
\arguments{
\item{x}{some data}
}
\value{
For Part One, \code{f07a(x)} returns .... For Part Two,
\code{f07b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/7}{The Treachery of Whales}
}
\details{
\strong{Part One}

A giant \href{https://en.wikipedia.org/wiki/Sperm_whale}{whale} has decided
your submarine is its next meal, and it\'s much faster than you are.
There\'s nowhere to run!

Suddenly, a swarm of crabs (each in its own tiny submarine - it\'s too
deep for them otherwise) zooms in to rescue you! They seem to be
preparing to blast a hole in the ocean floor; sensors indicate a
\emph{massive underground cave system} just beyond where they\'re aiming!

The crab submarines all need to be aligned before they\'ll have enough
power to blast a large enough hole for your submarine to get through.
However, it doesn\'t look like they\'ll be aligned before the whale
catches you! Maybe you can help?

There\'s one major catch - crab submarines can only move horizontally.

You quickly make a list of \emph{the horizontal position of each crab} (your
puzzle input). Crab submarines have limited fuel, so you need to find a
way to make all of their horizontal positions match while requiring them
to spend as little fuel as possible.

For example, consider the following horizontal positions:\preformatted{16,1,2,0,4,2,7,1,2,14
}

This means there\'s a crab with horizontal position \code{16}, a crab with
horizontal position \code{1}, and so on.

Each change of 1 step in horizontal position of a single crab costs 1
fuel. You could choose any horizontal position to align them all on, but
the one that costs the least fuel is horizontal position \code{2}:
\itemize{
\item Move from \code{16} to \code{2}: \code{14} fuel
\item Move from \code{1} to \code{2}: \code{1} fuel
\item Move from \code{2} to \code{2}: \code{0} fuel
\item Move from \code{0} to \code{2}: \code{2} fuel
\item Move from \code{4} to \code{2}: \code{2} fuel
\item Move from \code{2} to \code{2}: \code{0} fuel
\item Move from \code{7} to \code{2}: \code{5} fuel
\item Move from \code{1} to \code{2}: \code{1} fuel
\item Move from \code{2} to \code{2}: \code{0} fuel
\item Move from \code{14} to \code{2}: \code{12} fuel
}

This costs a total of \code{37} fuel. This is the cheapest possible outcome;
more expensive outcomes include aligning at position \code{1} (\code{41} fuel),
position \code{3} (\code{39} fuel), or position \code{10} (\code{71} fuel).

Determine the horizontal position that the crabs can align to using the
least fuel possible. \emph{How much fuel must they spend to align to that
position?}

\strong{Part Two}

\emph{(Use have to manually add this yourself.)}

\emph{(Try using \code{convert_clipboard_html_to_roxygen_md()})}
}
\examples{
f07a(input_07())
f07b(input_07())
}
