% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day22.R
\name{day22}
\alias{day22}
\alias{read_cubes}
\alias{reboot_reactor}
\title{Day 22}
\usage{
read_cubes(file)

reboot_reactor(reactor, init = FALSE)
}
\arguments{
\item{file}{A filename or text connection.}

\item{reactor}{A data frame as produced by \link{read_cubes}.}

\item{init}{Limit reboot to initialization procedure region?}
}
\description{
Day 22
}
