% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day23.R
\name{day23}
\alias{day23}
\alias{read_amphipod}
\alias{min_amphipod_energy}
\title{Day 23}
\usage{
read_amphipod(file)

min_amphipod_energy(start)
}
\arguments{
\item{file}{A filename or text connection.}

\item{start}{A \code{State} R6 object describing the initial position.}
}
\description{
Day 23
}
