% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_solution}
\alias{get_solution}
\title{Run solution script for a particular day}
\usage{
get_solution(day)
}
\arguments{
\item{day}{integer between 1 and 25}
}
\description{
Run solution script for a particular day
}
