% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day03.R
\name{most_common}
\alias{most_common}
\title{Most common binary digits}
\usage{
most_common(x)
}
\arguments{
\item{x}{A matrix or data frame of binary integers}
}
\value{
A binary vector of length \code{ncol(x)}
}
\description{
If 0 and 1 are equally common, 1 is returned (rounding up).
}
