% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.matrix}
\alias{read.matrix}
\title{Read a fixed-width matrix with no delimiter or header}
\usage{
read.matrix(file)
}
\arguments{
\item{file}{A text file or connection.}
}
\value{
A matrix.
}
\description{
Equivalent to \code{read.fwf} but faster and width does not need to be specified.
}
\examples{
read.matrix(input_file(3))

}
