% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{day02}
\alias{day02}
\alias{f02a}
\alias{f02b}
\alias{example_data_02}
\title{Day 02: Dive!}
\usage{
f02a(x)

f02b(x)

example_data_02(example = 1)
}
\arguments{
\item{x}{some data}

\item{example}{Which example data to use (by position or name). Defaults to
1.}
}
\value{
For Part One, \code{f02a(x)} returns .... For Part Two,
\code{f02b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/2}{Dive!}
}
\examples{
f02a(example_data_02())
f02b(example_data_02())
}
