% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day04.R
\name{day04}
\alias{day04}
\alias{f04a}
\alias{f04b}
\alias{get_numbers}
\alias{get_cards}
\title{Day 04: Giant Squid}
\usage{
f04a(numbers, cards)

f04b(numbers, cards)

get_numbers()

get_cards()
}
\arguments{
\item{numbers}{Numbers called so far}

\item{cards}{List of bingo cards}
}
\value{
For Part One, \code{f04a(x)} returns .... For Part Two,
\code{f04b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/4}{Giant Squid}
}
\examples{
f04a(get_numbers(), get_cards())
f04a(get_numbers(), get_cards())
}
