% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day07.R
\name{day07}
\alias{day07}
\alias{f07a}
\alias{f07b}
\title{Day 07: The Treachery of Whales}
\usage{
f07a(x)

f07b(x)
}
\arguments{
\item{x}{some data}
}
\value{
For Part One, \code{f07a(x)} returns .... For Part Two,
\code{f07b(x)} returns ....
}
\description{
\href{https://adventofcode.com/2021/day/7}{The Treachery of Whales}
}
\examples{
f07a(input_07())
f07b(input_07())
}
