% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day14.R
\name{day14}
\alias{day14}
\alias{Polymer}
\title{Day 14: Extended Polymerization}
\description{
Day 14: Extended Polymerization

Day 14: Extended Polymerization
}
\details{
\href{https://adventofcode.com/2021/day/14}{Extended Polymerization}
}
\examples{
\dontrun{
p <- Polymer$new()
p$take_steps(10)
p$get_answer()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{template}}{Starting template}

\item{\code{fun_call}}{Function call to \code{switch()} to convert two-letter pairs
according to rules}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Polymer-step}{\code{Polymer$step()}}
\item \href{#method-Polymer-take_steps}{\code{Polymer$take_steps()}}
\item \href{#method-Polymer-get_answer}{\code{Polymer$get_answer()}}
\item \href{#method-Polymer-clone}{\code{Polymer$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Polymer-step"></a>}}
\if{latex}{\out{\hypertarget{method-Polymer-step}{}}}
\subsection{Method \code{step()}}{
take one step
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Polymer$step()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Polymer-take_steps"></a>}}
\if{latex}{\out{\hypertarget{method-Polymer-take_steps}{}}}
\subsection{Method \code{take_steps()}}{
take n steps
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Polymer$take_steps(n = 10)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{number of steps to take}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Polymer-get_answer"></a>}}
\if{latex}{\out{\hypertarget{method-Polymer-get_answer}{}}}
\subsection{Method \code{get_answer()}}{
Calculate count of most common letter minus least common letter
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Polymer$get_answer()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Polymer-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Polymer-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Polymer$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
