% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/day02.R
\name{Submarine}
\alias{Submarine}
\title{Model the position of a submarine}
\description{
Model the position of a submarine

Model the position of a submarine
}
\examples{
dat <- "forward 5
down 5
forward 8
up 3
down 8
forward 2"
x <- strsplit(dat, "\n")[[1]]
s <- Submarine$new()
s$move(x)
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{position}}{horizontal position}

\item{\code{depth}}{vertical depth}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Submarine-forward}{\code{Submarine$forward()}}
\item \href{#method-Submarine-down}{\code{Submarine$down()}}
\item \href{#method-Submarine-up}{\code{Submarine$up()}}
\item \href{#method-Submarine-move}{\code{Submarine$move()}}
\item \href{#method-Submarine-clone}{\code{Submarine$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Submarine-forward"></a>}}
\if{latex}{\out{\hypertarget{method-Submarine-forward}{}}}
\subsection{Method \code{forward()}}{
Move forward
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Submarine$forward(x = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{forward distance to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Submarine-down"></a>}}
\if{latex}{\out{\hypertarget{method-Submarine-down}{}}}
\subsection{Method \code{down()}}{
Move down
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Submarine$down(x = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{depth to add}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Submarine-up"></a>}}
\if{latex}{\out{\hypertarget{method-Submarine-up}{}}}
\subsection{Method \code{up()}}{
Move up
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Submarine$up(x = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{depth to subtraxt}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Submarine-move"></a>}}
\if{latex}{\out{\hypertarget{method-Submarine-move}{}}}
\subsection{Method \code{move()}}{
Move based on a set of commands
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Submarine$move(commands)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{commands}}{vector of commands such as "forward 1", "down 2"}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Submarine-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Submarine-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Submarine$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
